New: build stream
