+++ stream/sink.go
@@ -1,33 +1,39 @@
-package std
+package stream
 
 import (
 	"github.com/BambooTuna/gooastream/queue"
-	"github.com/BambooTuna/gooastream/stream"
 )
 
 type (
 	sinkImpl struct {
 		in        queue.Queue
-		graphTree *stream.GraphTree
+		graphTree *GraphTree
 	}
 )
 
-var _ stream.Sink = (*sinkImpl)(nil)
+var _ Sink = (*sinkImpl)(nil)
 
-func NewSink(task func(interface{}) error, buffer int) stream.Sink {
+func BuildSink(in queue.Queue, graphTree *GraphTree) Sink {
+	return &sinkImpl{
+		in:        in,
+		graphTree: graphTree,
+	}
+}
+
+func NewSink(task func(interface{}) error, buffer int) Sink {
 	in := queue.NewQueueEmpty(buffer)
 	out := queue.NewQueueSink(task)
 	return &sinkImpl{
 		in:        in,
-		graphTree: stream.PassThrowGraph(in, out),
+		graphTree: PassThrowGraph(in, out),
 	}
 }
-func IgnoreSink() stream.Sink {
+func IgnoreSink() Sink {
 	in := queue.NewQueueEmpty(0)
 	out := queue.NewQueueSink(func(interface{}) error { return nil })
 	return &sinkImpl{
 		in:        in,
-		graphTree: stream.PassThrowGraph(in, out),
+		graphTree: PassThrowGraph(in, out),
 	}
 }
 func (a *sinkImpl) Dummy() {
@@ -35,6 +41,6 @@ func (a *sinkImpl) Dummy() {
 func (a *sinkImpl) In() queue.Queue {
 	return a.in
 }
-func (a *sinkImpl) GraphTree() *stream.GraphTree {
+func (a *sinkImpl) GraphTree() *GraphTree {
 	return a.graphTree
 }
